Creates a repository to each domain class
